+++ resources/mathjax-config.js
@@ -4,5 +4,23 @@
 
 window.MathJax = {
     showProcessingMessages: false,
-    messageStyle: "none"
+    messageStyle: "none",
+    skipStartupTypeset: true,
+    CommonHTML: {
+        scale: 90,
+        linebreaks: {
+            automatic: true
+        },
+        styles: {
+            ".MJXc-display": {
+                margin: "0"
+            }
+        }
+    },
+    "fast-preview": {
+        disabled: true
+    },
+    AuthorInit: function() {
+        MathJax.Hub.processSectionDelay = 0;
+    }
 };